Refactor: Extract generic hebcal helper functions to a separate file.
I will soon add an actually-shared function to this file.
